Add files via upload
This is e-commers webside . where we search product of e-commers , it will show a image and feature of that prouct . In this webside i am provide like and Add to card button . if you chick like button so that product go in  favorite page. if  we chick add to so go thise product in add to cart page. 

- I have use react to make this side.
- I have use functionl componets , hookes and contex api to manege state and data of componets  .
- To get data i am use js files.
- to manege seacher bar i am use debouncing .